requirements.txt
llama-parse
python-dotenv
langchain_community
langchain
chromadb
langchain-groq
chainlit